refactor(cli): drop duplicate promisify version
This is a cleanup after 57fe8589 and 4286c0d26, to ensure
all CLI source files are using the same promisify version
from utils.
